refactor(logic): extract deterministic engagement resolution (Refs #2288)
Move resolveEngagement and ratio/casualty helpers into combat_engagement.dart
to shrink combat_resolver.dart (~693 -> ~527 lines). Re-export
resolveEngagement from combat_resolver for unchanged barrel API.
Point combat_resolver_probabilistic at combat_types instead of importing
the full battle resolver.
