+++ packages/colonizethis_logic/lib/src/combat/combat_resolver.dart
@@ -11,13 +11,14 @@ import '../world/province_ownership_transfer.dart';
 import '../world/unit_lookup.dart';
 import 'battle_general_assignment.dart';
 import 'combat_constants.dart';
-import 'combat_loss_profile.dart';
+import 'combat_engagement.dart';
 import 'combat_types.dart';
 import 'conflict_detection.dart';
 import 'leader_bonus_helpers.dart';
 import 'military_strength.dart';
 
 export 'combat_constants.dart';
+export 'combat_engagement.dart' show resolveEngagement;
 export 'combat_types.dart';
 
 final _combatLog = packageLogger();
@@ -516,172 +517,6 @@ int _deploymentLimitForFaction(Game game, String factionId, int generalMedals) {
       ? deploymentLimitWithNationalism
       : deploymentLimitBase;
   return baseLimit + generalMedals;
-}
-
-/// Resolves one engagement: attacker vs defender.
-/// SPEC/program/combat-resolution.md.
-EngagementOutcome resolveEngagement({
-  required List<Unit> attackerUnits,
-  required List<Unit> defenderUnits,
-  int generalMedals = 0,
-  required int fortLevel,
-  required String terrain,
-  int defenderEffectiveMilitaryLevel = kDefaultEffectiveMilitaryEra,
-  double attackerMoraleMultiplier = kNeutralMoraleMultiplier,
-  double defenderMoraleMultiplier = kNeutralMoraleMultiplier,
-  double attackerLeaderMultiplier = kNeutralLeaderMultiplier,
-  double defenderLeaderMultiplier = kNeutralLeaderMultiplier,
-}) {
-  final attStr = aggregateStrength(attackerUnits, kDefaultEffectiveMilitaryEra);
-  var defStr = aggregateStrength(defenderUnits, defenderEffectiveMilitaryLevel);
-
-  final terrainMod =
-      terrainModifiers[terrain] ??
-      (kNeutralTerrainAttackerModifier, kNeutralTerrainDefenderModifier);
-  var effAtt =
-      attStr *
-      terrainMod.$1 *
-      attackerMoraleMultiplier *
-      attackerLeaderMultiplier;
-  var effDef =
-      defStr *
-      terrainMod.$2 *
-      defenderMoraleMultiplier *
-      defenderLeaderMultiplier;
-
-  if (fortLevel >= kMinFortLevelForCombatModifiers &&
-      fortLevel <= kMaxFortLevelForCombatModifiers) {
-    final reduction = fortDamageReduction[fortLevel];
-    effAtt *= (kUnityAttackerStrengthMultiplier - reduction);
-    final emplaced = fortGunCount[fortLevel] * fortEmplacedStrength[fortLevel];
-    effDef += emplaced;
-  }
-
-  // Wall HP soaks damage before it applies to defender casualty ratio. SPEC/game/siege-mechanics.md.
-  var effAttForRatio = effAtt;
-  if (fortLevel >= kMinFortLevelForCombatModifiers &&
-      fortLevel <= kMaxFortLevelForCombatModifiers) {
-    final wallHp = wallHpByFortLevel[fortLevel];
-    effAttForRatio = (effAtt - wallHp).clamp(
-      kEffectiveAttackForRatioClampMin,
-      kEffectiveAttackForRatioClampMax,
-    );
-  }
-
-  final attackerCasualties = <String>[];
-  final defenderCasualties = <String>[];
-
-  if (effAttForRatio <= 0 && effDef <= 0) {
-    return EngagementOutcome(
-      result: EngagementResult.stalemate,
-      attackerCasualties: attackerCasualties,
-      defenderCasualties: defenderCasualties,
-      attackerStrength: attStr,
-      defenderStrength: defStr,
-    );
-  }
-
-  final ratio = effDef > 0 ? effAttForRatio / effDef : kNoDefenderRatioFallback;
-  final attackerLowMorale = attackerMoraleMultiplier < defenderMoraleMultiplier;
-
-  return _resolveByRatio(
-    ratio: ratio,
-    attackerLowMorale: attackerLowMorale,
-    attackerUnits: attackerUnits,
-    defenderUnits: defenderUnits,
-    attStr: attStr,
-    defStr: defStr,
-  );
-}
-
-EngagementOutcome _resolveByRatio({
-  required double ratio,
-  required bool attackerLowMorale,
-  required List<Unit> attackerUnits,
-  required List<Unit> defenderUnits,
-  required double attStr,
-  required double defStr,
-}) {
-  final profile = combatLossProfileForStrengthRatio(
-    attackerDefenderStrengthRatio: ratio,
-    attackerLowMorale: attackerLowMorale,
-  );
-  return _buildOutcome(
-    attackerUnits: attackerUnits,
-    defenderUnits: defenderUnits,
-    attLossFrac: profile.attackerLossFraction,
-    defLossFrac: profile.defenderLossFraction,
-    attStr: attStr,
-    defStr: defStr,
-    bluntAttackerVictory: profile.bluntsAttackerVictory,
-    bothDeadResult: _engagementResultForMutualElimination(
-      profile.mutualEliminationOutcome,
-    ),
-  );
-}
-
-EngagementResult _engagementResultForMutualElimination(
-  CombatMutualEliminationOutcome outcome,
-) {
-  return switch (outcome) {
-    CombatMutualEliminationOutcome.attackerVictory =>
-      EngagementResult.attackerVictory,
-    CombatMutualEliminationOutcome.defenderVictory =>
-      EngagementResult.defenderVictory,
-    CombatMutualEliminationOutcome.mutualAnnihilation =>
-      EngagementResult.mutualAnnihilation,
-  };
-}
-
-EngagementOutcome _buildOutcome({
-  required List<Unit> attackerUnits,
-  required List<Unit> defenderUnits,
-  required double attLossFrac,
-  required double defLossFrac,
-  required double attStr,
-  required double defStr,
-  bool bluntAttackerVictory = false,
-  EngagementResult bothDeadResult = EngagementResult.mutualAnnihilation,
-}) {
-  final attLoss = combatCasualtyCount(
-    unitCount: attackerUnits.length,
-    lossFraction: attLossFrac,
-  );
-  final defLoss = combatCasualtyCount(
-    unitCount: defenderUnits.length,
-    lossFraction: defLossFrac,
-  );
-
-  final attackerCasualties = [
-    for (var i = 0; i < attLoss; i++) attackerUnits[i].id,
-  ];
-  final defenderCasualties = [
-    for (var i = 0; i < defLoss; i++) defenderUnits[i].id,
-  ];
-
-  final attSurvivors = attackerUnits.length - attLoss;
-  final defSurvivors = defenderUnits.length - defLoss;
-
-  EngagementResult result;
-  if (attSurvivors <= 0 && defSurvivors <= 0) {
-    result = bothDeadResult;
-  } else if (attSurvivors <= 0) {
-    result = EngagementResult.defenderVictory;
-  } else if (defSurvivors <= 0) {
-    result = bluntAttackerVictory
-        ? EngagementResult.stalemate
-        : EngagementResult.attackerVictory;
-  } else {
-    result = EngagementResult.stalemate;
-  }
-
-  return EngagementOutcome(
-    result: result,
-    attackerCasualties: attackerCasualties,
-    defenderCasualties: defenderCasualties,
-    attackerStrength: attStr,
-    defenderStrength: defStr,
-  );
 }
 
 /// Morale aura from general medals. SPEC/game/military-generals.md:
